refactor: address daily quality findings for 2026-08-06
Extract worktree inline-edit session/UI from ChangesCodeView, split sticky
agent attention latches into agent_hooks/attention.rs, and move terminal
cursor appearance settings into a dedicated settings group component.
